+++ vnext/ReactUWP/Modules/Animated/PropsAnimatedNode.cpp
@@ -234,6 +234,9 @@ void PropsAnimatedNode::MakeAnimation(int64_t valueNodeTag, FacadeType facadeTyp
         case FacadeType::Perspective:
           // TODO: implement perspective animations, tracked by issue #2680
           return;
+        case FacadeType::Progress:
+          // TODO: implement progress animations, tracked by issue #3283
+          return;
         default:
           assert(false);
       }