+++ vnext/ReactUWP/Views/Impl/SnapPointManagingContentControl.h
@@ -26,6 +26,9 @@ class SnapPointManagingContentControl
   // ScrollView Implementation
   void SnapToInterval(float interval);
   void SnapToOffsets(const winrt::IVectorView<float> &offsets);
+  void SnapToStart(bool snapToStart);
+  void SnapToEnd(bool snapToEnd);
+  void NotifySnapPointsUpdated();
 
   // IScrollSnapPointsInfo Implementation
   bool AreHorizontalSnapPointsRegular();
@@ -49,7 +52,7 @@ class SnapPointManagingContentControl
   void SetWidthBounds(float startWidth, float endWidth);
 
  private:
-  float m_interval{0};
+  float m_interval{0.0f};
   winrt::IVectorView<float> m_offsets{};
   bool m_snapToStart{true};
   bool m_snapToEnd{true};