+++ packages/react-native-windows-extended/CHANGELOG.md
@@ -1,7 +1,19 @@
 # Change Log - react-native-windows-extended
 
-This log was last generated on Sat, 26 Oct 2019 02:55:31 GMT and should not be manually modified.
+This log was last generated on Fri, 01 Nov 2019 23:40:12 GMT and should not be manually modified.
 
+## 0.60.15
+Fri, 01 Nov 2019 23:40:12 GMT
+
+### Patches
+
+- Updating react-native to version: 0.60.0-microsoft.15 (53619745+rnbot@users.noreply.github.com)
+## 0.60.14
+Thu, 31 Oct 2019 18:11:20 GMT
+
+### Patches
+
+- Updating react-native to version: 0.60.0-microsoft.14 (53619745+rnbot@users.noreply.github.com)
 ## 0.60.13
 Sat, 26 Oct 2019 02:55:31 GMT